apenas um enpoint está sendo utilizado para testes. objetos estão sendo criados no controler por enquanto. classes model foram organizadas para estarem de acordo com o banco de dados

+++ src/main/java/com/example/loja/database/ProductsInterface.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductsInterface extends JpaRepository<Product, Long> {
-
-
+public interface ProductsInterface extends JpaRepository<Product, Integer> {
 
 }
